stt_api/requirements.txt
Flask
deepmultilingualpunctuation
openai>=1.0.0
python-dotenv